QPID-7739 : [Java Broker] [AMQP 1.0] Use correct types in operations with performative fields
* use int for channel id
* avoid setting of negative maxFrameSize (enforce maxFrameSize to be less than networkBufferSize)
* in comparison operations for Disposition#first and Disposition#last use SequenceNumber

+++ broker-core/src/main/java/org/apache/qpid/server/model/Connection.java
@@ -91,7 +91,7 @@ public interface Connection<X extends Connection<X>> extends ConfiguredObject<X>
     String getRemoteProcessPid();
 
     @DerivedAttribute
-    long getSessionCountLimit();
+    int getSessionCountLimit();
 
     @DerivedAttribute
     Transport getTransport();
