+++ broker-plugins/amqp-0-10-protocol/src/main/java/org/apache/qpid/server/protocol/v0_10/AMQPConnection_0_10Impl.java
@@ -335,7 +335,7 @@ public class AMQPConnection_0_10Impl extends AbstractAMQPConnection<AMQPConnecti
     }
 
     @Override
-    public long getSessionCountLimit()
+    public int getSessionCountLimit()
     {
         return _connection.getSessionCountLimit();
     }
