+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/framing/SASLFrame.java
@@ -29,7 +29,7 @@ public final class SASLFrame extends AMQFrame<SaslFrameBody>
         super(frameBody);
     }
 
-    @Override public short getChannel()
+    @Override public int getChannel()
     {
         return (short)0;
     }
