+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/type/security/SaslMechanisms.java
@@ -67,7 +67,7 @@ public class SaslMechanisms
         return builder.toString();
     }
 
-    public void invoke(final short channel, SASLEndpoint conn)
+    public void invoke(final int channel, SASLEndpoint conn)
     {
         conn.receiveSaslMechanisms(this);
     }
