+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/type/security/SaslOutcome.java
@@ -88,7 +88,7 @@ public class SaslOutcome
         return builder.toString();
     }
 
-    public void invoke(final short channel, SASLEndpoint conn)
+    public void invoke(final int channel, SASLEndpoint conn)
     {
         conn.receiveSaslOutcome(this);
     }
