+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/type/security/SaslResponse.java
@@ -66,7 +66,7 @@ public class SaslResponse
         return builder.toString();
     }
 
-    public void invoke(final short channel, SASLEndpoint conn)
+    public void invoke(final int channel, SASLEndpoint conn)
     {
         conn.receiveSaslResponse(this);
     }
