+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/type/transport/Close.java
@@ -65,7 +65,7 @@ public class Close implements FrameBody
         return builder.toString();
     }
 
-    public void invoke(short channel, ConnectionHandler conn)
+    public void invoke(int channel, ConnectionHandler conn)
     {
         conn.receiveClose(channel, this);
     }
